+++ tests/data/oci.sql
@@ -326,3 +326,29 @@ CREATE TABLE "bit_values" (
 );
 
 INSERT INTO "bit_values" ("id", "val") VALUES (1, '0'), (2, '1');
+
+
+/* viatable test, see https://github.com/yiisoft/yii2/issues/5004 */
+
+BEGIN EXECUTE IMMEDIATE 'DROP TABLE "collection"'; EXCEPTION WHEN OTHERS THEN IF SQLCODE != -942 THEN RAISE; END IF; END;--
+BEGIN EXECUTE IMMEDIATE 'DROP TABLE "collection_item"'; EXCEPTION WHEN OTHERS THEN IF SQLCODE != -942 THEN RAISE; END IF; END;--
+
+CREATE TABLE "collection" (
+  "id" integer NOT NULL,
+  "version" integer NOT NULL,
+  CONSTRAINT "collection_PK" PRIMARY KEY ("id", "version") ENABLE
+);
+
+INSERT INTO "collection" ("id", "version") VALUES (1, 1);
+
+CREATE TABLE "collection_item" (
+  "collection_id" integer NOT NULL,
+  "collection_version" integer NOT NULL,
+  "item_id" integer NOT NULL,
+  CONSTRAINT "collection_type_PK" PRIMARY KEY ("collection_id", "collection_version", "item_id") ENABLE
+);
+
+INSERT INTO "collection_item" ("collection_id", "collection_version", "item_id") VALUES (1, 1, 1);
+INSERT INTO "collection_item" ("collection_id", "collection_version", "item_id") VALUES (1, 1, 2);
+INSERT INTO "collection_item" ("collection_id", "collection_version", "item_id") VALUES (1, 1, 3);
+